+++ src/mergeSortClass.h
@@ -14,20 +14,20 @@ public:
 	void cleanCnt();							// 清除若干计数器
 	
 	double run_action(); 						// 配置不同的数据并启动排序
-	void action(long long int*, int, int);		// mergeSort 的实现,主函数
-	void merge(long long int*, int, int, int);	// 归并函数
+	void action(ll*, ll, ll);		// mergeSort 的实现,主函数
+	void merge(ll*, ll, ll, ll);	// 归并函数
 	
 	dataGenerator* getDataGenPtr();				// 获取dataGenPtr
-	long long int getCompareCnt();				// 获取compareCnt
-	long long int getExchageCnt();				// 获取exchageCnt
+	ll getCompareCnt();				// 获取compareCnt
+	ll getExchageCnt();				// 获取exchageCnt
 	
 	mergeSortClass(testData* data);				// 构造
 	~mergeSortClass();							// 析构
 private:
 	dataGenerator* dataGenPtr;
-	long long int* assistantArr;
-	long long int compareCnt;
-	long long int exchageCnt;
+	ll* assistantArr;
+	ll compareCnt;
+	ll exchageCnt;
 };
 
 #endif
